discord_chat: handle strange out of memory issue
Server stop responsing while there is a out of memory message in the log.
Server still have planty of memory, but this issue happens around
every 2 days, and server is still responsing query from battlemetrics, but
player can't find it on the game server list.
Restart the server if we found this message in the log.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>